+++ docs/fftl.rst
@@ -1,1 +1,8 @@
-.. automodule:: fftl
+Generalised FFTLog
+==================
+
+The functionality of the :mod:`fftl` module is provided by the
+:func:`fftl.transform` routine to compute the generalised FFTLog integral
+transform for a given kernel.
+
+.. autofunction:: fftl.transform
