+++ docs/transforms.rst
@@ -1,1 +1,10 @@
-.. automodule:: fftl.transforms
+Standard Integral Transforms
+============================
+
+The :mod:`fftl` module provides Python implementations for a number of standard
+integral transforms.
+
+.. autoclass:: fftl.HankelTransform
+.. autoclass:: fftl.LaplaceTransform
+.. autoclass:: fftl.SphericalHankelTransform
+.. autoclass:: fftl.StieltjesTransform
